requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
aniso8601==8.1.1
bidict==0.21.2
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
dnspython==1.16.0
eventlet==0.30.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.2
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2021.1
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
